requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
bcrypt==3.2.2
black==26.3.0
boto3==1.42.63
botocore==1.42.63
cffi==2.0.0
cfgv==3.5.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
distlib==0.4.0
ecdsa==0.19.1
fastapi==0.133.1
filelock==3.25.0
flake8==7.3.0
greenlet==3.3.2
h11==0.16.0
httptools==0.7.1
identify==2.6.17
idna==3.11
isort==8.0.1
jmespath==1.1.0
librt==0.8.1
Mako==1.3.10
MarkupSafe==3.0.3
mccabe==0.7.0
mypy==1.19.1
mypy-boto3-s3==1.42.37
mypy_extensions==1.1.0
nodeenv==1.10.0
packaging==26.0
passlib==1.7.4
pathspec==1.0.4
platformdirs==4.9.4
pre_commit==4.5.1
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pyflakes==3.4.0
python-dateutil==2.9.0.post0
python-discovery==1.1.1
python-dotenv==1.2.2
python-jose==3.5.0
python-json-logger==4.0.0
python-multipart==0.0.22
pytokens==0.4.1
PyYAML==6.0.3
rsa==4.9.1
ruff==0.15.5
s3transfer==0.16.0
six==1.17.0
SQLAlchemy==2.0.48
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
virtualenv==21.1.0
watchfiles==1.1.1
websockets==16.0
